+++ CHANGELOG.md
@@ -60,6 +60,7 @@ Cinecred Changelog
 ### UI Fixes
 
 - The preview indicator for embedded videos is now high-res and less obnoxious.
+- The color picker now allows dragging the brightness even when in HDR mode.
 - The video preview now updates the duration when switching between sheets.
 - The video preview no longer retains an afterimage upon exiting full screen.
 - In the video preview, fades and transparent embedded videos now play smoothly.